인풋수정및 스텟SO

+++ Assets/00.Work/SW/01.Scripts/Player/StatSO.cs
@@ -9,5 +9,5 @@ public class StatSO : ScriptableObject
     public int barrierCount;
     public int damage;
     public int curBlletCount;
-    public int cooltime;
+    public float cooltime;
 }
